output_projects/New-Project/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
sqlalchemy==2.0.31
aiosqlite==0.20.0
python-multipart==0.0.9
python-jose[cryptography]==3.9.0
passlib[bcrypt]==1.7.4
cryptography==42.0.8
httpx==0.27.0
openai==1.35.3
google-generativeai==0.7.2
groq==0.9.0
aiofiles==23.2.1
jinja2==3.1.4
pydub==0.25.1
websockets==12.0